+++ recipes/setup_server.rb
@@ -15,13 +15,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# Define the service with a custom name so we can subscribe just to the "restart" action
-# otherwise the "wait_for_webapp" resource will be notified too early (when tomcat is stopped)
-# and enqueued before the restart action is triggered
-service "abiquo-tomcat-start" do
-    service_name "abiquo-tomcat"
-end
-
 template "/var/www/html/ui/config/client-config-custom.json" do
     source "ui-config.json.erb"
     owner "root"
@@ -32,31 +25,7 @@ template "/var/www/html/ui/config/client-config-custom.json" do
         :ui_address => node['abiquo']['ui_address']
     })
     action :create
+    notifies :restart, "service[apache2]"
 end
 
-template "/opt/abiquo/tomcat/conf/server.xml" do
-    source "server.xml.erb"
-    owner "root"
-    group "root"
-    action :create
-    notifies :start, "service[abiquo-tomcat-start]"
-end
-
-template "/opt/abiquo/config/abiquo.properties" do
-    source "abiquo.properties.erb"
-    owner "root"
-    group "root"
-    variables(lazy {{ :properties => node['abiquo']['properties'] }})
-    action :create
-    notifies :restart, "service[abiquo-tomcat-start]"
-end
-
-abiquo_wait_for_webapp "api" do
-    host "localhost"
-    port node['abiquo']['tomcat']['http-port']
-    retries 3   # Retry if Tomcat is still not started
-    retry_delay 5
-    action :nothing
-    subscribes :wait, "service[abiquo-tomcat-start]"
-    only_if { node['abiquo']['tomcat']['wait-for-webapps'] }
-end
+include_recipe "abiquo::service"