+++ spec/repository_spec.rb
@@ -19,24 +19,24 @@ describe 'abiquo::repository' do
 
     it 'cleans the yum cache' do
         chef_run.converge(described_recipe)
-        expect(chef_run).to run_execute('clean-yum-cache').with(
-            :command => 'yum clean all'
-        )
+        resource = chef_run.execute('clean-yum-cache')
+        expect(resource).to do_nothing
+        expect(resource.command).to eq('yum clean all')
     end
 
     it 'deletes the yum cache directory' do
         chef_run.converge(described_recipe)
-        expect(chef_run).to delete_directory('/var/cache/yum').with(
-            :ignore_failure => true,
-            :recursive => true
-        )
+        resource = chef_run.directory('/var/cache/yum')
+        expect(resource).to do_nothing
+        expect(resource.ignore_failure).to eq(true)
+        expect(resource.recursive).to eq(true)
     end
 
     it 'creates the base repository' do
         chef_run.converge(described_recipe)
         expect(chef_run).to create_yum_repository('abiquo-base').with(
             :description => 'Abiquo base repository',
-            :baseurl => 'http://mirror.abiquo.com/abiquo/3.6/os/x86_64',
+            :baseurl => 'http://mirror.abiquo.com/abiquo/3.8/os/x86_64',
             :gpgcheck => true,
             :gpgkey => 'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-Abiquo ' \
                        'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-MariaDB ' \
@@ -46,14 +46,16 @@ describe 'abiquo::repository' do
         )
 
         resource = chef_run.find_resource(:yum_repository, 'abiquo-base')
-        expect(resource).to subscribe_to('package[abiquo-release-ee]').on(:create)
+expect(resource).to subscribe_to('package[abiquo-release-ee]').on(:create)
+        expect(resource).to notify('directory[/var/cache/yum]').to(:delete).immediately
+        expect(resource).to notify('execute[clean-yum-cache]').to(:run).immediately
     end
 
     it 'creates the updates repository' do
         chef_run.converge(described_recipe)
         expect(chef_run).to create_yum_repository('abiquo-updates').with(
             :description => 'Abiquo updates repository',
-            :baseurl => 'http://mirror.abiquo.com/abiquo/3.6/updates/x86_64',
+            :baseurl => 'http://mirror.abiquo.com/abiquo/3.8/updates/x86_64',
             :gpgcheck => true,
             :gpgkey => 'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-Abiquo ' \
                        'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-MariaDB ' \
@@ -64,6 +66,8 @@ describe 'abiquo::repository' do
 
         resource = chef_run.find_resource(:yum_repository, 'abiquo-updates')
         expect(resource).to subscribe_to('package[abiquo-release-ee]')
+        expect(resource).to notify('directory[/var/cache/yum]').to(:delete).immediately
+        expect(resource).to notify('execute[clean-yum-cache]').to(:run).immediately
     end
 
     it 'creates does not create the nightly packages repository by default' do
@@ -84,6 +88,9 @@ describe 'abiquo::repository' do
                        'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-CentOS-6 ' \
                        'file:///etc/pki/rpm-gpg/RPM-GPG-RSA-KEY-Abiquo'
         )
+        resource = chef_run.find_resource(:yum_repository, 'abiquo-nightly')
+        expect(resource).to notify('directory[/var/cache/yum]').to(:delete).immediately
+        expect(resource).to notify('execute[clean-yum-cache]').to(:run).immediately
     end
 
     it 'installs the abiquo-release-ee package' do
@@ -91,5 +98,10 @@ describe 'abiquo::repository' do
         expect(chef_run).to install_package('abiquo-release-ee').with(
             :options => '--nogpgcheck'
         )
+    end
+
+    it 'installs yum-utils package' do
+        chef_run.converge(described_recipe)
+        expect(chef_run).to install_package('yum-utils')
     end
 end