+++ spec/support/matchers.rb
@@ -16,10 +16,6 @@ def import_java_management_truststore_certificate(resource_name)
     ChefSpec::Matchers::ResourceMatcher.new(:java_management_truststore_certificate, :import, resource_name)
 end
 
-def dump_ark(resource_name)
-    ChefSpec::Matchers::ResourceMatcher.new(:ark, :dump, resource_name)
-end
-
 def permissive_selinux_state(resource_name)
     ChefSpec::Matchers::ResourceMatcher.new(:selinux_state, :permissive, resource_name)
 end