+++ spec/upgrade_spec.rb
@@ -13,14 +13,56 @@
 # limitations under the License.
 
 require 'spec_helper'
+require_relative 'support/packages'
 
 describe 'abiquo::upgrade' do
-    let(:chef_run) { ChefSpec::SoloRunner.new }
+    let(:chef_run) { ChefSpec::SoloRunner.new(internal_locale: 'en_US.UTF-8') }
+
+    before do
+        stub_command('/usr/sbin/httpd -t').and_return(true)
+        stub_command("service abiquo-tomcat stop").and_return(true)
+        allow(::File).to receive(:executable?).with('/usr/bin/repoquery').and_return(true)
+        stub_package_commands(['abiquo-api', 'abiquo-server'])
+    end
+
+    it 'does nothing if repoquery is not installed' do
+        allow(::File).to receive(:executable?).with('/usr/bin/repoquery').and_return(false)
+        chef_run.converge('apache2::default', described_recipe)
+        expect(chef_run.find_resource(:service, 'abiquo-tomcat')).to be_nil
+        expect(chef_run.find_resource(:service, 'abiquo-aim')).to be_nil
+        expect(chef_run.find_resource(:package, 'abiquo-api')).to be_nil
+    end
+
+    it 'logs a message if there are upgrades' do
+        chef_run.converge('apache2::default', described_recipe)
+        expect(chef_run).to write_log("Abiquo updates available.")
+    end
+
+    it 'logs a message if there are no upgrades' do
+        stub_available_packages(['abiquo-api', 'abiquo-server'], '-0:3.6.1-85.el6.noarch')
+        chef_run.converge('apache2::default', described_recipe)
+        expect(chef_run).to write_log("No Abiquo updates found.")
+    end
+
+    it 'does nothing if no updates available' do
+        stub_available_packages(['abiquo-api', 'abiquo-server'], '-0:3.6.1-85.el6.noarch')
+        chef_run.converge('apache2::default', described_recipe)
+        expect(chef_run.find_resource(:service, 'abiquo-tomcat')).to be_nil
+        expect(chef_run.find_resource(:service, 'abiquo-aim')).to be_nil
+        expect(chef_run.find_resource(:package, 'abiquo-api')).to be_nil
+    end
+
+    it 'performs upgrade if there are new rpms' do
+        chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
+        expect(chef_run.find_resource(:service, 'abiquo-tomcat')).not_to be_nil
+        expect(chef_run.find_resource(:service, 'abiquo-aim')).to be_nil
+        expect(chef_run.find_resource(:package, 'abiquo-api')).not_to be_nil
+    end
 
     %w{monolithic remoteservices server v2v}.each do |profile|
         it "stops the #{profile} services" do
             chef_run.node.set['abiquo']['profile'] = profile
-            chef_run.converge(described_recipe)
+            chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
             expect(chef_run).to stop_service('abiquo-tomcat')
             expect(chef_run).to_not stop_service('abiquo-aim')
             expect(chef_run).to_not stop_service('abiquo-delorean')
@@ -30,7 +72,7 @@ describe 'abiquo::upgrade' do
 
     it 'stops the kvm services' do
         chef_run.node.set['abiquo']['profile'] = 'kvm'
-        chef_run.converge(described_recipe)
+        chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
         expect(chef_run).to_not stop_service('abiquo-tomcat')
         expect(chef_run).to_not stop_service('abiquo-delorean')
         expect(chef_run).to_not stop_service('abiquo-emmett')
@@ -39,7 +81,7 @@ describe 'abiquo::upgrade' do
 
     it 'stops the monitoring services' do
         chef_run.node.set['abiquo']['profile'] = 'monitoring'
-        chef_run.converge(described_recipe)
+        chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
         expect(chef_run).to_not stop_service('abiquo-tomcat')
         expect(chef_run).to_not stop_service('abiquo-aim')
         expect(chef_run).to stop_service('abiquo-delorean')
@@ -47,21 +89,56 @@ describe 'abiquo::upgrade' do
     end
 
     it 'includes the repository recipe' do
-        chef_run.converge(described_recipe)
+        chef_run.converge('apache2::default', described_recipe)
         expect(chef_run).to include_recipe('abiquo::repository')
     end
 
-    it 'upgrades the abiquo packages' do
-        chef_run.converge(described_recipe)
-        expect(chef_run).to run_execute('yum-upgrade-abiquo').with(
-            :command => 'yum -y upgrade abiquo-*'
-        )
+    it 'upgrades the abiquo packages on monolithic' do
+        chef_run.converge('apache2::default',described_recipe)
+        %w{abiquo-api abiquo-server}.each do |pkg|
+            expect(chef_run).to upgrade_package(pkg)
+        end
+    end
+
+    it 'upgrades the abiquo packages on server' do
+        chef_run.node.set['abiquo']['profile'] = 'server'
+        chef_run.converge('apache2::default',described_recipe)
+        %w{abiquo-api abiquo-server}.each do |pkg|
+            expect(chef_run).to upgrade_package(pkg)
+        end
+    end
+
+    it 'upgrades the abiquo packages on remoteservices' do
+        stub_package_commands(['abiquo-virtualfactory', 'abiquo-remote-services'])
+        chef_run.node.set['abiquo']['profile'] = 'remoteservices'
+        chef_run.converge('apache2::default', described_recipe)
+        %w{abiquo-virtualfactory abiquo-remote-services}.each do |pkg|
+            expect(chef_run).to upgrade_package(pkg)
+        end
+    end
+
+    it 'upgrades the abiquo packages on kvm' do
+        stub_package_commands(['abiquo-aim', 'abiquo-cloud-node'])
+        chef_run.node.set['abiquo']['profile'] = 'kvm'
+        chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
+        %w{abiquo-aim abiquo-cloud-node}.each do |pkg|
+            expect(chef_run).to upgrade_package(pkg)
+        end
+    end
+
+    it 'upgrades the abiquo packages on monitoring' do
+        stub_package_commands(['abiquo-delorean', 'abiquo-emmett'])
+        chef_run.node.set['abiquo']['profile'] = 'monitoring'
+        chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
+        %w{abiquo-delorean abiquo-emmett}.each do |pkg|
+            expect(chef_run).to upgrade_package(pkg)
+        end
     end
 
     it 'does not run liquibase if not configured' do
         chef_run.node.set['abiquo']['db']['upgrade'] = false
         chef_run.node.set['abiquo']['profile'] = 'monolithic'
-        chef_run.converge(described_recipe)
+        chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
         expect(chef_run).to_not run_execute('liquibase-update')
     end
 
@@ -69,57 +146,66 @@ describe 'abiquo::upgrade' do
         it "does not run liquibase when #{profile}" do
             chef_run.node.set['abiquo']['db']['upgrade'] = true
             chef_run.node.set['abiquo']['profile'] = profile
-            chef_run.converge(described_recipe)
+            chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
             expect(chef_run).to_not run_execute('liquibase-update')
         end
     end
 
     it 'runs the liquibase update when monolithic' do
-        chef_run.converge(described_recipe)
-        expect(chef_run).to run_execute('liquibase-update').with(
-            :cwd => '/usr/share/doc/abiquo-server/database',
-            :command => 'abiquo-liquibase -h localhost -P 3306 -u root update'
-        )
+        chef_run.converge('apache2::default', described_recipe)
+        resource = chef_run.find_resource(:execute, 'liquibase-update')
+        expect(resource).to subscribe_to("package[abiquo-server]").on(:run).immediately
+        expect(resource).to do_nothing
+        expect(resource.cwd).to eq('/usr/share/doc/abiquo-server/database')
+        expect(resource.command).to eq('abiquo-liquibase -h localhost -P 3306 -u root update')
+        expect(resource).to notify('service[abiquo-tomcat]').to(:restart).delayed
     end
 
     it 'runs the liquibase update with custom attributes' do
         chef_run.node.set['abiquo']['db']['host'] = '127.0.0.1'
         chef_run.node.set['abiquo']['db']['password'] = 'abiquo'
-        chef_run.converge(described_recipe)
-        expect(chef_run).to run_execute('liquibase-update').with(
-            :cwd => '/usr/share/doc/abiquo-server/database',
-            :command => 'abiquo-liquibase -h 127.0.0.1 -P 3306 -u root -p abiquo update'
-        )
+        chef_run.converge('apache2::default', described_recipe)
+        resource = chef_run.find_resource(:execute, 'liquibase-update')
+        expect(resource).to subscribe_to("package[abiquo-server]").on(:run).immediately
+        expect(resource).to do_nothing
+        expect(resource.cwd).to eq('/usr/share/doc/abiquo-server/database')
+        expect(resource.command).to eq('abiquo-liquibase -h 127.0.0.1 -P 3306 -u root -p abiquo update')
+        expect(resource).to notify('service[abiquo-tomcat]').to(:restart).delayed
     end
 
     %w{monolithic remoteservices server v2v}.each do |profile|
-        it "notifies the #{profile} service to restart" do
+        it "starts the #{profile} services" do
             chef_run.node.set['abiquo']['profile'] = profile
-            chef_run.converge(described_recipe)
-            resource = chef_run.execute('yum-upgrade-abiquo')
-            expect(resource).to notify('service[abiquo-tomcat]').to(:start).delayed
+            chef_run.converge('apache2::default', described_recipe)
+            expect(chef_run).to start_service('abiquo-tomcat')
+            expect(chef_run).to_not start_service('abiquo-aim')
+            expect(chef_run).to_not start_service('abiquo-delorean')
+            expect(chef_run).to_not start_service('abiquo-emmett')
         end
     end
 
-    it 'notifies the kvm service to restart' do
+    it 'starts the kvm services' do
         chef_run.node.set['abiquo']['profile'] = 'kvm'
-        chef_run.converge(described_recipe)
-        resource = chef_run.execute('yum-upgrade-abiquo')
-        expect(resource).to notify('service[abiquo-aim]').to(:start).delayed
+        chef_run.converge('apache2::default', described_recipe)
+        expect(chef_run).to_not start_service('abiquo-tomcat')
+        expect(chef_run).to_not start_service('abiquo-delorean')
+        expect(chef_run).to_not start_service('abiquo-emmett')
+        expect(chef_run).to start_service('abiquo-aim')
     end
 
-    it 'notifies the monitoring services to restart' do
+    it 'starts the monitoring services' do
         chef_run.node.set['abiquo']['profile'] = 'monitoring'
-        chef_run.converge(described_recipe)
-        resource = chef_run.execute('yum-upgrade-abiquo')
-        expect(resource).to notify('service[abiquo-delorean]').to(:start).delayed
-        expect(resource).to notify('service[abiquo-emmett]').to(:start).delayed
+        chef_run.converge('apache2::default', described_recipe)
+        expect(chef_run).to_not start_service('abiquo-tomcat')
+        expect(chef_run).to_not start_service('abiquo-aim')
+        expect(chef_run).to start_service('abiquo-delorean')
+        expect(chef_run).to start_service('abiquo-emmett')
     end
 
     %w(monolithic remoteservices kvm monitoring server v2v).each do |profile|
         it "includes the #{profile} setup recipe" do
             chef_run.node.set['abiquo']['profile'] = profile
-            chef_run.converge(described_recipe)
+            chef_run.converge('apache2::default', described_recipe, 'abiquo::service')
             expect(chef_run).to include_recipe("abiquo::setup_#{profile}")
         end
     end