+++ test/integration/monolithic/serverspec/monolithic_config_spec.rb
@@ -34,14 +34,14 @@ describe 'Monolithic configuration' do
         expect(file('/etc/httpd/sites-available/abiquo.conf')).to contain('SSLEngine on')
         expect(file('/etc/httpd/sites-available/abiquo.conf')).to contain('SSLProtocol all -SSLv2')
         expect(file('/etc/httpd/sites-available/abiquo.conf')).to contain('SSLCipherSuite ALL:!ADH:!EXPORT:!SSLv2:RC4+RSA:+HIGH:+MEDIUM:+LOW')
-        expect(file('/etc/httpd/sites-available/abiquo.conf')).to contain('SSLCertificateFile /usr/var/ssl/certs/server.crt')
-        expect(file('/etc/httpd/sites-available/abiquo.conf')).to contain('SSLCertificateKeyFile /usr/var/ssl/certs/server.key')
+        expect(file('/etc/httpd/sites-available/abiquo.conf')).to contain('SSLCertificateFile /etc/pki/abiquo/monolithic.abiquo.com.crt')
+        expect(file('/etc/httpd/sites-available/abiquo.conf')).to contain('SSLCertificateKeyFile /etc/pki/abiquo/monolithic.abiquo.com.key')
     end
 
     it 'has the ui properly configured' do
         expect(file('/var/www/html/ui/config/client-config-custom.json')).to exist
         # By default the uri will get the hostname (generated from the name of the suite)
-        expect(file('/var/www/html/ui/config/client-config-custom.json')).to contain('"config.endpoint": "https://monolithic-centos-65/api"')
+        expect(file('/var/www/html/ui/config/client-config-custom.json')).to contain('"config.endpoint": "https://monolithic.abiquo.com/api"')
     end
 
     it 'has tomcat properly configured' do