+++ test/integration/v2v/serverspec/v2v_packages_spec.rb
@@ -27,6 +27,15 @@ describe 'Remote Services packages' do
         end
     end
 
+    it 'has the iscsi-initiator-utils package installed' do
+        expect(package('iscsi-initiator-utils')).to be_installed
+    end
+
+    it 'has the iSCSI initiator name configured' do
+        expect(file('/etc/iscsi/initiatorname.iscsi')).to exist
+        expect(file('/etc/iscsi/initiatorname.iscsi').size).to be > 0
+    end
+
     it 'has the strong jce encryption policies installed' do
         expect(file('/usr/java/default/jre/lib/security/local_policy.jar').md5sum).to eq('dabfcb23d7bf9bf5a201c3f6ea9bfb2c')
         expect(file('/usr/java/default/jre/lib/security/US_export_policy.jar').md5sum).to eq('ef6e8eae7d1876d7f05d765d2c2e0529')